add answer policy

+++ app/controllers/answers_controller.rb
@@ -11,6 +11,7 @@ class AnswersController < ApplicationController
   end
 
   def create
+    authorize Answer
     @answer = question.answers.new(answer_params)
     @answer.author = current_user
     @answer.save
@@ -22,6 +23,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
+    authorize answer
     if current_user.author_of?(answer)
       answer.destroy
       flash[:notice] = t('.success')
@@ -31,6 +33,7 @@ class AnswersController < ApplicationController
   end
 
   def mark_as_the_best
+    authorize answer
     @answer = Answer.find(params[:best_answer])
     @question = Question.find(params[:id])
     @answer.set_best_answer(@question) if current_user.author_of?(@question)
